+++ optics-th/CHANGELOG.md
@@ -1,4 +1,5 @@
-# optics-th-0.4.2 (2026-??-??)
+# optics-th-0.5 (2026-??-??)
+* Drop support for GHC older than 9.2.
 * Require `optics-core >= 0.5`.
 
 # optics-th-0.4.1 (2022-03-22)